+++ feature/exercise-chart/src/main/kotlin/io/github/stslex/workeeper/feature/exercise_chart/mvi/mapper/ExerciseChartUiMapper.kt
@@ -24,6 +24,7 @@ internal object ExerciseChartUiMapper {
     fun ChartPointDomain.toUi(): ChartPointUiModel = ChartPointUiModel(
         day = day,
         dayMillis = dayMillis,
+        sessionUuid = sessionUuid,
         value = value,
         setCount = setCount,
     )